+++ synapse-compaction-aws-s3/src/main/java/de/otto/synapse/compaction/s3/CompactionService.java
@@ -47,6 +47,7 @@ public class CompactionService {
         this.messageLogReceiverEndpointFactory = messageLogReceiverEndpointFactory;
         this.clock = clock;
     }
+
     public String compact(final String channelName, final MessageFormat messageFormat) {
         LOG.info("Start compacting channel {}", channelName);
         stateRepository.clear();